feat(cta/split-button): add basic changes
