+++ react-examples/stories/CallsToAction/ButtonGroup.stories.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import TwixtButtonGroup from '../../react/src/CallsToAction/ButtonGroup';
-import TwixtButton from '../../react/src/CallsToAction/Button';
-import TwixtIcon from '../../react/src/Icon';
+import TwixtButtonGroup from '../../../react/src/CallsToAction/ButtonGroup';
+import TwixtButton from '../../../react/src/CallsToAction/Button';
+import TwixtIcon from '../../../react/src/Icon';
 import { background } from '@storybook/theming';
 
 export default {
